controllerのdestory修正

+++ app/controllers/items_controller.rb
@@ -38,8 +38,10 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if (user_signed_in? && current_user.id == item.user.id)
+      item.destroy
+      redirect_to root_path
+    end
   end
 
   private
